+    - FEATURE: Now remembers filtered pieces in a torrent on restart +    - FEATURE: Now updating pieces progress in real time in torrent properties +    - BUGFIX: Fixed memory leaks in torrent properties Updated italian and German translation

+++ src/downloadThread.h
@@ -119,7 +119,7 @@ class downloadThread : public QThread {
             QFile::remove(filePath.c_str());
             return;
           }
-	  int retries = 0;
+	  unsigned short retries = 0;
 	  bool to_many_users = false;
 	  do{
 	    // Perform Download
